Replace utils/pool with promise pool, use limit 4 for parse
Moved the old src/utils/pool.ts to src/utils/thread-pool.ts since it's
just a wrapper for threads.js's Pool. Opened up src/utils/pool.ts to be
@connormanning's more useful Promise pool utility, which I used in place
of Promise.all in src/utils/parse.ts, as suggested.

Other changes are files that previously imported src/utils/pool.ts
directly.
