+++ src/utils/index.ts
@@ -5,4 +5,5 @@ export * from './parse.js'
 export * from './pool.js'
 export * from './status.js'
 export * from './testing.js'
+export * from './thread-pool.js'
 export * from './vlrs.js'
